Edited README for structure and content
Hey--per Matt's request, I've gone through and added some structure and standard content sections to the README, to bring it closer in line to what's in the README's of the other published Graph SDK projects. Feel free to edit/add/delete the content as it applies to the Ruby project. Let me know if you have any questions. Thanks.